main update
