+++ frontend/src/Components/Yehoon/Components/HomePage/MainArticle/MainArticle.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Theme from "./../../../../../theme/theme";
+import { Theme } from "./../../../../../theme/theme";
 import { StyledButton } from "./../../../../../theme/theme";
 
 const MainArticleBox = styled.div`
