+++ frontend/src/Components/Yehoon/Components/HomePage/MainLatestProduct/LatestProductItem.js
@@ -1,5 +1,6 @@
 import styled from "styled-components";
-import Theme from "./../../../../../theme/theme";
+import { Theme } from "./../../../../../theme/theme";
+import { Link } from "react-router-dom";
 
 const LatestProductItemBox = styled.div`
   width: 180px;
@@ -13,40 +14,62 @@ const LatestProductItemBox = styled.div`
     cursor: pointer;
   }
 
-  & > img {
-    width: 163px;
-    height: 168px;
-  }
+  & a {
+    &:link {
+      text-decoration: none;
+      color: black;
+    }
+    &:visited {
+      text-decoration: none;
+      color: black;
+    }
+    &:active {
+      text-decoration: none;
+      color: black;
+    }
 
-  & > div:nth-child(2) {
-    & > div:nth-child(1) {
-      font-family: "SCD-6";
-      font-size: 1.7rem;
+    &:hover {
+      cursor: pointer;
       color: ${Theme.green};
     }
 
-    & > div:nth-child(2) {
-      font-family: "SCD-3";
-      font-size: 1.6rem;
-      color: rgba(0, 0, 0, 1);
+    & > img {
+      width: 163px;
+      height: 168px;
     }
-  }
-  & > div:nth-child(3) {
-    font-family: "SCD-7";
-    font-size: 1.9rem;
-    color: ${Theme.lightblack};
+
+    & > div:nth-child(2) {
+      & > div:nth-child(1) {
+        font-family: "SCD-6";
+        font-size: 1.7rem;
+        color: ${Theme.green};
+      }
+
+      & > div:nth-child(2) {
+        font-family: "SCD-3";
+        font-size: 1.6rem;
+        color: ${Theme.lightblack};
+      }
+    }
+    & > div:nth-child(3) {
+      font-family: "SCD-7";
+      font-size: 1.9rem;
+      color: ${Theme.lightblack};
+    }
   }
 `;
 
 const LatestProductItem = ({ data }) => {
   return (
     <LatestProductItemBox>
-      <img src={data.image} alt="이미지 없음"></img>
-      <div>
-        <div>[{data.local}]</div>
-        <div>{data.name}</div>
-      </div>
-      <div>{data.price}</div>
+      <Link to={`/product/${data.product_id}`}>
+        <img src={data.image} alt="이미지 없음"></img>
+        <div>
+          <div>[{data.local}]</div>
+          <div>{data.name}</div>
+        </div>
+        <div>{data.price}</div>
+      </Link>
     </LatestProductItemBox>
   );
 };
