+++ frontend/src/Components/Yehoon/Components/ProductPage/ProductCategory.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const CategoryBox = styled.div`
+  width: 1280px;
+  height: 28px;
+`;
+
+const ProductCategory = () => {
+  return (
+    <CategoryBox>
+      <div>하이</div>
+    </CategoryBox>
+  );
+};
+
+export default ProductCategory;
